src/ru/job4j/array/SortSelectedTest.java
package ru.job4j.array;

import org.junit.Test;
import static org.hamcrest.core.Is.is;
import static org.junit.Assert.assertThat;

public class SortSelectedTest {
    @Test
    public void whenSort() {
        int[] input = new int[] {3, 4, 1, 2, 5};
        int[] result = SortSelected.sort(input);
        int[] expect = new int[] {1, 2, 3, 4, 5};
        assertThat(result, is(expect));
    }
    @Test
    public void whenSort1() {
        int[] input = new int[] {3, 4, 1};
        int[] result = SortSelected.sort(input);
        int[] expect = new int[] {1, 3, 4};
        assertThat(result, is(expect));
    }
    @Test
    public void whenSort2() {
        int[] input = new int[] {56, 45, 9, 32, 15};
        int[] result = SortSelected.sort(input);
        int[] expect = new int[] {9, 56, 32, 45, 15};
        assertThat(result, is(expect));
    }
    @Test
    public void whenSort3() {
        int[] input = new int[] {56, 56, 9, 32, 15};
        int[] result = SortSelected.sort(input);
        int[] expect = new int[] {9, 56, 32, 45, 15};
        assertThat(result, is(expect));
    }
}
